Code for hillslope model
This is the first iteration of my code for my EV320 Final Project. This code is able to produce 2 initial elevation plots. One plot is in 3D and the other is a colormesh plot. The current issue with this code is that the final elevation plots are not being produced. The running time for the project is still very low, 1000 years, but this is being used to compute the model at quicker speeds.