Final Code before Annotations
I added a color map of the gradient/slope on both 3D figures so it is easier to read the terrain. I also removed the colormesh figures from the output as they are hard to read without much context. Finally, I changed the run time from 1000 years to 250 years. I did this to accurately display the effects of hiking trails on a hillslope landscape because, after 1000 years, the hiking trails are essentially erased from the landscape.